Overloaded basic math funcs (double and float variants). Use real_t rather than float or double in generic functions (core/math) whenever possible.
Also inlined some more math functions.

+++ scene/gui/range.cpp
@@ -141,8 +141,8 @@ void Range::set_as_ratio(double p_value) {
 
 	if (shared->exp_ratio && get_min()>0) {
 
-		double exp_min = Math::log(get_min())/Math::log(2);
-		double exp_max = Math::log(get_max())/Math::log(2);
+		double exp_min = Math::log(get_min())/Math::log((double)2);
+		double exp_max = Math::log(get_max())/Math::log((double)2);
 		v = Math::pow(2,exp_min+(exp_max-exp_min)*p_value);
 	} else {
 
@@ -160,9 +160,9 @@ double Range::get_as_ratio() const {
 
 	if (shared->exp_ratio && get_min()>0) {
 
-		double exp_min = Math::log(get_min())/Math::log(2);
-		double exp_max = Math::log(get_max())/Math::log(2);
-		double v = Math::log(get_value())/Math::log(2);
+		double exp_min = Math::log(get_min())/Math::log((double)2);
+		double exp_max = Math::log(get_max())/Math::log((double)2);
+		double v = Math::log(get_value())/Math::log((double)2);
 
 		return (v - exp_min) / (exp_max - exp_min);
